relocate 'get_shortest_path_*()' to `Utilities` submodule
